Split encoders and decoders into separate files.
