--- acnportal/io/__init__.py
@@ -1,1 +1,0 @@
-from .json.json_io import from_json, to_json
